# docs/phase5/logs/openapi-v1-phase5.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "DerasaX.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/Units/GetUnitsBySubjectId": {
      "get": {
        "tags": [
          "Units"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Units/GetUnitsBySubjectId": {
      "get": {
        "tags": [
          "Units"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/Units/AddUnit": {
      "post": {
        "tags": [
          "Units"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "Title": {
                    "type": "string"
                  },
                  "SubjectId": {
                    "type": "string"
                  }
                }
              },
              "encoding": {
                "Title": {
                  "style": "form"
                },
                "SubjectId": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Units/AddUnit": {
      "post": {
        "tags": [
          "Units"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "Title": {
                    "type": "string"
                  },
                  "SubjectId": {
                    "type": "string"
                  }
                }
              },
              "encoding": {
                "Title": {
                  "style": "form"
                },
                "SubjectId": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/Units/UpdateUnit": {
      "put": {
        "tags": [
          "Units"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "Id": {
                    "type": "string"
                  },
                  "Title": {
                    "type": "string"
                  },
                  "SubjectId": {
                    "type": "string"
                  }
                }
              },
              "encoding": {
                "Id": {
                  "style": "form"
                },
                "Title": {
                  "style": "form"
                },
                "SubjectId": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Units/UpdateUnit": {
      "put": {
        "tags": [
          "Units"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "Id": {
                    "type": "string"
                  },
                  "Title": {
                    "type": "string"
                  },
                  "SubjectId": {
                    "type": "string"
                  }
                }
              },
              "encoding": {
                "Id": {
                  "style": "form"
                },
                "Title": {
                  "style": "form"
                },
                "SubjectId": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/Units/DeleteUnit": {
      "delete": {
        "tags": [
          "Units"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Units/DeleteUnit": {
      "delete": {
        "tags": [
          "Units"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
...
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}